fix linting issue

+++ core/services/gateway/handlers/webapicapabilities/webapi.go
@@ -11,7 +11,7 @@ type TargetRequestPayload struct {
 type TargetResponsePayload struct {
 	ExecutionError bool              `json:"executionError"`         // true if there were non-HTTP errors. false if HTTP request was sent regardless of status (2xx, 4xx, 5xx)
 	ErrorMessage   string            `json:"errorMessage,omitempty"` // error message in case of failure
-	StatusCode     uint16            `json:"statusCode,omitempty"`   // HTTP status code
+	StatusCode     int               `json:"statusCode,omitempty"`   // HTTP status code
 	Headers        map[string]string `json:"headers,omitempty"`      // HTTP headers
 	Body           []byte            `json:"body,omitempty"`         // HTTP response body
 }
